add http client/handler
